pyproject.toml
[tool.poetry]
name = "pyawg"
version = "0.1.0"
description = ""
authors = ["Abhishek Bawkar <abhishekbawkar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
python-vxi11 = "^0.9"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
